Fix URL Encoded path being passed during search
At some point the filename was being URL escaped which caused the regex looking for an ID to fail. This was causing some super weird behavior, such as episode titles having the first "a" in them replaced with a space, and all other metadata failing to be found. Adding this one line fixes that entirely, though it may cause hiccups if a file for whatever reason HAS url escaped characters in the name on purpose.