Fixed missing Director
When the Setting "Set YouTube usernames as director in metadata" is enabled and the file `info.json` file is available then the director should be set to the `channel_title`. This results in the director being empty because the `channel_title` is never set (except initially with an empty string).

I added that the director is being filled from the uploader element from the info.json.