Fix issue where agent ignores channel_id and playlist_id in path where local JSON file if present
Agent was ignoring defined channel_id and playlist_id in path if JSON file was present. This would result in the video_id being used as the guid and metadata for the playlist or channel not being properly scraped. The above reordering still allows this method to be used if these values are not present in path.